move html-producting functionality to `html` module
