+++ src/formatter/markdown.rs
@@ -10,9 +10,10 @@ use regex::Regex;
 use walkdir::WalkDir;
 
 use crate::config::Config;
-use crate::formatter::highlight;
-use crate::formatter::utils::{escape_html, generate_id};
-use crate::template;
+use crate::html::{
+    highlight, template,
+    utils::{escape_html, generate_id},
+};
 
 /// Represents a header in the markdown document
 #[derive(Debug, Clone, serde::Serialize)]
@@ -207,9 +208,9 @@ pub fn process_markdown(
 fn preprocess_inline_anchors(content: &str) -> String {
     // First handle list items with anchors at the beginning
     let mut result = String::with_capacity(content.len() + 100);
-    let mut lines = content.lines().peekable();
+    let lines = content.lines().peekable();
 
-    while let Some(line) = lines.next() {
+    for line in lines {
         if let Some(caps) = LIST_ITEM_WITH_ANCHOR_RE.captures(line) {
             let list_marker = caps.get(1).unwrap().as_str();
             let id = caps.get(2).unwrap().as_str();
@@ -719,7 +720,7 @@ fn process_empty_auto_links(html: String) -> String {
         .to_string()
 }
 
-/// Process empty HTML links (as generated by pulldown_cmark)
+/// Process empty HTML links (as generated by `pulldown_cmark`)
 fn process_html_empty_links(html: String) -> String {
     HTML_EMPTY_LINK_RE
         .replace_all(&html, |caps: &regex::Captures| {
