+++ src/formatter/mod.rs
@@ -1,5 +1,3 @@
-pub mod highlight;
 pub mod manpage;
 pub mod markdown;
 pub mod options;
-mod utils;
